moved plottools to new visualization folder, added plot_vispy to it too
