+++ openpnm/topotools/__init__.py
@@ -10,4 +10,3 @@ network topology.
 from ._topotools import *
 from ._perctools import *
 from ._graphtools import *
-from ._plottools import *
